+++ packages/components/cypress/e2e/popovercontainer.cy.ts
@@ -1,12 +1,12 @@
-describe('popovercontainer', () => {
+describe('popovercontainer', { baseUrl: null, includeShadowDom: true }, () => {
   describe('default', () => {
     let selector = isPopoverSupported() ? ':popover-open' : '.\\:popover-open';
 
     beforeEach(() => {
       // There is no dedicated docs page for the popovercontainer
-      cy.getComponent('popover');
+      cy.visit('./cypress/fixtures/post-popover.test.html');
       cy.get('[data-popover-target="popover-one"][aria-expanded]').as('trigger');
-      cy.get('post-popovercontainer').as('container');
+      cy.get('div.popover-container').as('container');
     });
 
     it('should show up on click', () => {