+++ packages/documentation/.storybook/constants.ts
@@ -3,4 +3,6 @@ export const BADGE = {
   STABLE: 'stable',
   NEEDS_REVISION: 'needs-revision',
   TODO: 'to-do',
+  SINCE_V1: 'Since v1',
+  WEB_COMPONENT_CANDIDATE: 'Web Component Candidate',
 };