+++ packages/internet-header/cypress/e2e/meta-navigation.cy.ts
@@ -4,21 +4,22 @@ import { prepare } from '../support/prepare-story';
 
 describe('meta-navigation', () => {
   beforeEach(() => {
-    prepare('Internet Header/Header', 'Default');
+    prepare('Components/Internet Header/Header', 'Default');
     cy.changeArg('language', 'de');
     cy.viewport(1024, Cypress.config('viewportHeight'));
   });
 
   describe('args', () => {
     describe('meta: true', () => {
-      it(`adds meta navigation`, () => {
-        cy.changeArg('meta', true);
+      it.only(`adds meta navigation`, () => {
+        cy.changeArg('meta', 'true');
+        cy.pause();
         cy.get('post-meta-navigation').should('exist').and('be.visible');
       });
 
       it(`sets css variable --meta-header-height to 0 in mobile view`, () => {
         cy.viewport(1023, Cypress.config('viewportHeight'));
-        cy.changeArg('meta', true);
+        cy.changeArg('meta', 'true');
         cy.get('swisspost-internet-header').within(() => {
           cy.get('header').then($el => {
             const style = getComputedStyle($el[0]);
@@ -30,7 +31,7 @@ describe('meta-navigation', () => {
 
       it(`sets css variable --meta-header-height to 3rem in desktop view`, () => {
         cy.viewport(1024, Cypress.config('viewportHeight'));
-        cy.changeArg('meta', true);
+        cy.changeArg('meta', 'true');
         cy.get('swisspost-internet-header').within(() => {
           cy.get('header').then($el => {
             const style = getComputedStyle($el[0]);
@@ -43,12 +44,12 @@ describe('meta-navigation', () => {
 
     describe('meta: false', () => {
       it(`removes meta navigation`, () => {
-        cy.changeArg('meta', false);
+        cy.changeArg('meta', 'false');
         cy.get('post-meta-navigation').should('not.exist');
       });
 
       it(`sets css variable --meta-header-height to 0`, () => {
-        cy.changeArg('meta', false);
+        cy.changeArg('meta', 'false');
         cy.get('swisspost-internet-header').within(() => {
           cy.get('header').then($el => {
             const style = getComputedStyle($el[0]);
@@ -61,13 +62,13 @@ describe('meta-navigation', () => {
 
     describe('meta: changes during runtime (false => true => false)', () => {
       it(`adds and removes meta navigation`, () => {
-        cy.changeArg('meta', false);
+        cy.changeArg('meta', 'false');
         cy.get('post-meta-navigation').should('not.exist');
 
-        cy.changeArg('meta', true);
+        cy.changeArg('meta', 'true');
         cy.get('post-meta-navigation').should('exist').and('be.visible');
 
-        cy.changeArg('meta', false);
+        cy.changeArg('meta', 'false');
         cy.get('post-meta-navigation').should('not.exist');
       });
     });
@@ -96,7 +97,7 @@ describe('meta-navigation', () => {
         // Clear meta navigation config
         config.de!.header.navMeta = undefined;
 
-        prepare('Internet Header/Header', 'Default', { config });
+        prepare('Components/Internet Header/Header', 'Default', { config });
         cy.changeArg('language', 'de');
 
         // Assert the header is hydrated
@@ -122,7 +123,7 @@ describe('meta-navigation', () => {
           },
         ];
 
-        prepare('Internet Header/Header', 'Default', { config });
+        prepare('Components/Internet Header/Header', 'Default', { config });
         cy.changeArg('language', 'de');
 
         // Assert the header is hydrated