+++ packages/components-react/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @swisspost/design-system-components-react
 
+## 1.0.26
+
+### Patch Changes
+
+- Updated dependencies:
+  - @swisspost/design-system-components@2.0.0
+
+## 1.0.25
+
+### Patch Changes
+
+- Updated dependencies:
+  - @swisspost/design-system-components@1.7.1
+
+## 1.0.24
+
+### Patch Changes
+
+- Updated dependencies:
+  - @swisspost/design-system-components@1.7.0
+
+## 1.0.23
+
+### Patch Changes
+
+- Updated dependencies:
+  - @swisspost/design-system-components@1.6.3
+
 ## 1.0.22
 
 ### Patch Changes