+++ packages/intranet-header-workspace/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @swisspost/design-system-intranet-header-workspace
 
+## 3.0.7
+
+### Patch Changes
+
+- Updated dependencies:
+  - @swisspost/design-system-styles@6.6.0
+
+## 3.0.6
+
+### Patch Changes
+
+- Updated dependencies:
+  - @swisspost/design-system-styles@6.5.1
+
+## 3.0.5
+
+### Patch Changes
+
+- Updated dependencies:
+  - @swisspost/design-system-styles@6.5.0
+
 ## 3.0.4
 
 ### Patch Changes